Add files via upload
Let's Start with Stream lit Library function is open source tool used to build  & share interactive web apps..
This Stream lit code creates a simple web app that displays an image titled "Display Image Example." It uses the PIL (Python Imaging Library) to open an image file named "solve.jpg". The image is then displayed on the app with the caption "Sunset in the mountains", and use_column_width=200 adjusts the image width to fit within the specified column size. This code allows users to view the image with a customized layout and caption in a Stream lit app.